Multi-user functionality working, Multi-room not.
One giant commit for the day since I temp forgot how git works (oops).

Chatroom has been added again for debug purposes. It will be removed once we have everything off the ground. At the moment, the movie list is non-randomised and is in order (design to be discussed later). Users within the same room will vote, when everybody says Yes to a film a popup will come up but nothing more at the moment.

This version is BROKEN for multi-room - i.e. room/24 and room/348 will be merged (they all have to say "yes" to a certain film and they all have the same users). Will be fixed in the near future.

Users functionality is broken, and signing in is no longer present (for test purposes only). Just go straight to /room/ to log in.

Default pictures will have to be changed.

This works by (on initalisation), grabbing 20 movies from the API and adding them to a list on the server. Each client connected to a room keeps an index into this list and tells the server if they have voted "Yes". The yes_count of that film is incremented and if it is equal to the number of users on the server then a film has been found. (All of the above is broken for multi-room).

Contributors:
Chase Hellemans (ch3213)
Henry Cuttell (hc3313)