Added scrolling functionality to the chatroom
Small internal style css to the chatroom to allow it to scroll
(and auto-scroll whenever there is an update), just for ease.

Also upped the buffer from 5 to 10 with the consent of Henry,
as we found out that if you press films too quickly, the index
becomes out of sync as the next set of films are not ready to use
because the server has not fetched them yet. Just to see what happens from here.