Added post-confirm redirection and room teardown
When all users have left, the room gets torn down so all the yes_votes
and movie list is new whenever anybody enters that room; as if all traces
of that room have gone, allowing us to re-use room IDs.

Confirms have been changed to Alerts, and they redirect the user elsewhere.
That elsewhere is undecided and is currently now the empty ID room.