Added Admin and Force-Go functionality.
When a user creates a room, they become the admin (for that room).
Only they can see the Go button in the lobby and those
who join (and are therefore not the admin) cannot.
This is done by hiding the Go button when you join,
and showing it when you Create

Once the admin hits "Go", then all the users (including
the admin), will be brought to the main page.