Found film redirects to room page instead of first. User list bug fixes.
After a movie has been found, a the page is redirected to the Room page
instead of the first page, allowing for users to remain logged in,
only their room number has become unset.

Fixed the user list such that you can also see the users in the
lobby and the film pages. A small bug on updating the list upon
disconnect has been done.