Film Page doesn't load until all films fetched.
The code has changed a lot for this one.

generate_films has a callback function to
make all users to move on to the film page
as soon as all the information for the movie
(incl. the extra data) has been fetched.

This causes it to lag a little bit however
'waiting_signal' allows you to make a feedback
notification to all users that the room is being built
As of now it is just adding "BUILDING" to the room
number.

Also added shuffling of films as they are added, to
add a little more randomisation to the list of films