Added multi-user preferences. Also genre bug fix
When 'go' is pressed, users send their preferences
to the server. Once the server has all the votes
(genre_collection_count vs num_users), only then does
the server generate the films and move the users
to the film page.

There was a slight bug with genres, "TV Movie" was being
passed through as just "TV", causing the ID translation
to be undefined. I've added underscores instead of whitespace
to avoid this, however now the lobby page looks bad.
Needs to be fixed in the future