Fixes to run server on multiple computers
No longer listens to localhost exclusively.

To run this server on multiple computers, simply run it as usual
(node server.js), and then go on to your localhost:8080.
You will neeed to find our outgoing IP address, for Linux, use
'ifconfig' and look at the network device you are using (wlan0 for
wireless, eth0 for wired). Others computers will use this as an address
and 8080 as the port (e.g. 146.98.123.23:8080). Check the console log
to see if they have connected properly.