Genre tickbox list only occurs on page load, Leave Room Transition change
Because genre_add_tickboxes appends the html file, it should definitely
only happen once and not whenever the socket connects (previously
was in socket_connect). Now it has been moved to document.ready(),
which occurs only once whenever the page loads. Since this is a
single-page application, this is fine.