Added time out to room teardown to prevent bug
Room tears down 10 seconds after a film being found
to prevent the server crashing until a solution can be
found.

The bug occurs when a request for more films gets sent
out and does not return before the room has been
torn down, causing the function to add to a deleted list.