Fixed bug in free_resources that tore down other rooms
free_resources() tore down the user's current room.
However if this was called after the user had made a new
room, it will tear down the new room.

I realise that this is a byproduct of the previous commit
that makes room teardown happen after 10 seconds, however
since we can expect room teardown to no longer happen
instantaneously, this was a needed fix.

Username and room is now passed into free_resources,
ensuring it tears down the correct room.